add: error page, checkout

+++ store/src/components/CartProduct.tsx
@@ -2,6 +2,7 @@ import { useState, useContext } from "react"
 import { Button, Container, Navbar, Modal, Nav } from "react-bootstrap"
 import { CartContext } from "../CartContext"
 import { getProductData } from "../productStore"
+import NotFound from "./404"
 
 function CartProduct(props: any) {
   const cart = useContext(CartContext)
@@ -9,17 +10,21 @@ function CartProduct(props: any) {
   const quantity = props.quantity
   const productData = getProductData(id)
 
-  return (
-    <>
-      <h3>{productData?.title}</h3>
-      <p>{quantity} total</p>
-      <p>${(quantity * productData?.price).toFixed(2)}</p>
-      <Button size="sm" onClick={() => cart.deleteFromCart(id)}>
-        Remove
-      </Button>
-      <hr></hr>
-    </>
-  )
+  if (productData?.price !== undefined) {
+    return (
+      <>
+        <h3>{productData?.title}</h3>
+        <p>{quantity} total</p>
+        <p>${(quantity * productData?.price).toFixed(2)}</p>
+        <Button size="sm" onClick={() => cart.deleteFromCart(id)}>
+          Remove
+        </Button>
+        <hr></hr>
+      </>
+    )
+  } else {
+    return <NotFound />
+  }
 }
 
 export default CartProduct
